backend/requirements.txt
sanic==23.3.0
sanic-ext==23.3.0
sanic-testing==23.3.0
sqlalchemy==2.0.3
asyncpg==0.27.0
Jinja2==3.1.2
PyJWT==2.6.0
bcrypt==4.0.1
pydantic==1.10.7
geopy==2.3.0
pytest==7.2.1
pytest-asyncio==0.20.3
google==3.0.0
google-auth==2.16.2
openai==0.27.0
aioredis==2.0.1
motor==3.1.1
pymongo==4.3.3
boto3==1.26.109
Pillow==9.5.0
scikit-learn==1.2.2
Faker==18.10.0
opennsfw2==0.10.2
